requirements.txt
tensorflow==2.0.0
flask==1.1.1
cassandra-driver==3.22.0
Pillow==7.0.0